move report summary graphs, add scrolling
