+++ src/app/report-rollup/assessment-reports/assessment-reports.component.html
@@ -1,5 +1,8 @@
 <!-- NOT PRINT -->
 <div *ngIf="!printView">
+    <div class="assessment-item" id="{{'rollup_graphs'}}">
+        <app-report-summary-graphs class="w-100"></app-report-summary-graphs>
+    </div>
     <div *ngFor="let item of _psatAssessments" id="{{'assessment_'+item.assessment.id}}" class="assessment-item">
         <app-psat-report *ngIf="item.assessment.psat" [assessment]="item.assessment" [inRollup]="true"
             [settings]="item.settings">
