+++ src/app/report-rollup/report-summary-graphs/report-summary-graphs.component.css
