+++ src/app/report-rollup/report-summary-graphs/report-summary-pie-chart/report-summary-pie-chart.component.html
@@ -1,0 +1,1 @@
+<div class="w-100" #reportSummaryPieChart></div>
