--- src/app/report-rollup/report-summary/report-summary-graphs/report-summary-graphs.component.html
@@ -1,20 +1,0 @@
-<div class="row no-gutters">
-        <div class="col-12 d-flex justify-content-center">
-                <div class="energy-pie-chart" *ngIf="pieChartData">
-                        <app-report-summary-pie-chart class="chart" [pieChartData]="pieChartData"
-                                [dataOption]="pieChartDataOption" [energyUnit]="reportSummaryEnergyUnit"
-                                [titleStr]="'Assessment Energy Usage Summary'">
-                        </app-report-summary-pie-chart>
-                </div>
-
-
-                <div class="energy-pie-chart" *ngIf="energyChartData">
-                        <app-report-summary-pie-chart class="chart" [pieChartData]="energyChartData"
-                                [dataOption]="pieChartDataOption" [energyUnit]="reportSummaryEnergyUnit"
-                                [titleStr]="'Fuel vs Electricity Energy Usage'">
-                        </app-report-summary-pie-chart>
-                </div>
-
-
-        </div>
-</div>
