--- src/app/report-rollup/report-summary/report-summary-graphs/report-summary-pie-chart/report-summary-pie-chart.component.html
@@ -1,1 +1,0 @@
-<div #reportSummaryPieChart></div>
