+++ src/app/report-rollup/report-summary/report-summary.component.html
@@ -45,8 +45,8 @@
     </div>
 
 
-    <app-report-summary-graphs
+    <!-- <app-report-summary-graphs
         *ngIf="showPsatSummary || showPhastSummary || showFsatSummary || showSsmtSummary || showWasteWater || showCompressedAir"
         [hidden]="showSummary == 'closed'">
-    </app-report-summary-graphs>
+    </app-report-summary-graphs> -->
 </div>
